frontend/src/app/modules/product/pages/product-read/read.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { NO_ERRORS_SCHEMA } from '@angular/core';
import { ProductService } from '../../product.service';
import { MatDialog } from '@angular/material/dialog';
import { RouterTestingModule } from '@angular/router/testing';
import { ReadComponent } from './read.component';

describe('ReadComponent', () => {
  let component: ReadComponent;
  let fixture: ComponentFixture<ReadComponent>;

  beforeEach(() => {
    const productServiceStub = () => ({
      read: () => ({ subscribe: f => f({}) })
    });
    const matDialogStub = () => ({
      open: (productDeleteComponent, object) => ({})
    });
    TestBed.configureTestingModule({
      imports: [RouterTestingModule],
      schemas: [NO_ERRORS_SCHEMA],
      declarations: [ReadComponent],
      providers: [
        { provide: ProductService, useFactory: productServiceStub },
        { provide: MatDialog, useFactory: matDialogStub }
      ]
    });
    fixture = TestBed.createComponent(ReadComponent);
    component = fixture.componentInstance;
  });

  it('can load instance', () => {
    expect(component).toBeTruthy();
  });

  it(`displayedColumns has default value`, () => {
    expect(component.displayedColumns).toEqual([
      `id`,
      `name`,
      `price`,
      `action`
    ]);
  });

  describe('ngOnInit', () => {
    it('makes expected calls', () => {
      const productServiceStub: ProductService = fixture.debugElement.injector.get(
        ProductService
      );
      spyOn(productServiceStub, 'read').and.callThrough();
      component.ngOnInit();
      expect(productServiceStub.read).toHaveBeenCalled();
    });
  });

  describe('openDeleteModal', () => {
    it('makes expected calls', () => {
      const matDialogStub: MatDialog = fixture.debugElement.injector.get(
        MatDialog
      );
      spyOn(matDialogStub, 'open').and.callThrough();
      component.openDeleteModal();
      expect(matDialogStub.open).toHaveBeenCalled();
    });
  });
});
